toInt to toDouble

+++ dotnet grpc client/Client/Program.cs
@@ -133,7 +133,7 @@ class Program
                 System.Console.WriteLine("Please enter height of plane:");
                 double hg = Convert.ToDouble(Console.ReadLine());
                 System.Console.WriteLine("Please enter weight of plane:");
-                double wg = Convert.ToInt32(Console.ReadLine());
+                double wg = Convert.ToDouble(Console.ReadLine());
                 var pl = new PlaneProto(nm,float.Parse(hg.ToString()),float.Parse(wg.ToString()));
                 var addTask = client.AddPlane(pl.toAddRequest());
                 System.Console.WriteLine($"server responsed :{addTask.RespMessage}");
